Refactor form [TEA-21]
