+++ packages/client/src/components/Form/index.ts
@@ -1,6 +1,6 @@
 export { type FormProps, Form } from './Form';
 export {
   useFormContext,
-  type FormSubmitHandler,
   type FormContextState,
-} from './FormContext';
+} from './context';
+export { type FormSubmitHandler } from './types';
